fix update_rule and add partial spec
