+++ kensu/utils/rule_engine.py
@@ -1,5 +1,7 @@
 from kensu.utils.kensu_provider import KensuProvider
 import warnings
+import logging
+
 
 
 def add_rule(data_source, field, type, parameters,context="DATA_STATS"):
@@ -9,6 +11,25 @@ def add_rule(data_source, field, type, parameters,context="DATA_STATS"):
                          'fun': {"name": type, "arguments": parameters},
                          'context':context
                          }})
+
+
+def add_not_null_rule(
+        ksu,  # type: Kensu
+        lds_name,
+        non_null_col,
+        null_suffix='nullrows'
+):
+    try:
+        logging.info(f"Adding a Kensu rule: NOT_NULL({non_null_col}) on LDS={lds_name}")
+        lds_context = "LOGICAL_DATA_SOURCE"
+        add_rule(data_source=lds_name,
+                 field=f'{non_null_col}.{null_suffix}',
+                 type='Range',
+                 parameters={'maxVal': 0},
+                 context=lds_context)
+        ksu.send_rules()
+    except Exception as e:
+        logging.error(f"Error while adding a Kensu rule  NOT_NULL({non_null_col}.{null_suffix})", e)
 
 
 def add_min_max(data_source, field, min = None, max = None, context="DATA_STATS"):
